Remove next/prev proposal buttons
Before the sidebar list of proposals was available on the proposal
detail page, we (devs) gave ourselves next/previous buttons for quickly
navigating to new proposals. (It was never offered to production users
because it simply tried to go to adjacent proposals by incrementing or
decrementing the sequential ID, regardless of whether there actually
*was* a proposal with that ID. This was an acceptable limitation for dev
use, but not for production use.)

Now that we have the sidebar, we don’t need these anymore.

Issue #172 Remove next/previous proposal dev buttons