Allow text selection on proposal list table
Prior to this commit, text selection of contents of the proposal list
table wasn’t possible. Clicking, dragging, and releasing the cursor
triggered the `click` event, which caused us to navigate to the proposal
detail page.

This commit modifies the click handler so that we only navigate if there
is no text selected. (This uses the unintuitive `isCollapsed` property
of the selection object, which is `true` when the selection’s start and
end positions are the same — i.e., nothing is selected.)

Issue #85 Click handler prevents text selection on proposal table