Forbid Dependabot from upgrading Storybook
Storybook has a bunch of components to choose from. Those components
should be updated alongside each other; mixing different versions is not
a good practice, and ends up causing a lot of PR thrash and email spam.

The team working on Dependabot is considering a feature to allow for
grouped updates, where we'd be able to specify that all of the Storybook
components should be updated at the same time, rather than separate PRs.
Until that feature is implemented, though, tell Dependabot to ignore
Storybook.