Link to RECENT_CHANGES.md from landing page
Adds a link to the `RECENT_CHANGES.md` document on our GitHub to the
landing page. Note that this links to the `main` branch version of the
file, which as of this moment doesn’t exist, but will once this PR that
contains this commit is merged. Schrödinger’s Link, or something.

Issue #39 Add changelog / What's New page