Remove user-facing references to “data viewer”
This project will be expanded to support sending data into the PDC, so
it is no longer just a data *viewer*.

Since it is essentially the canonical view into the PDC for most people
(short of the API), we can simply refer to it as “the Philanthropy Data
Commons” in most user-facing language.

While we are in the pilot phase, we will continue to append the
word “pilot”, but this commit rewords all uses so that the term “data
viewer” doesn’t appear and so the word “pilot” can be removed in the
future.

Issue #315 Choose new name for Data Viewer