Create skeleton of add-data page
Build out the route, an initial page, and a few placeholder components
for the new Add Data page, and wire the placeholder components into
Storybook.

There are still several open questions that this does not attempt to
resolve, including:

- what information do we get from the API about previous bulk uploads?
- what exactly do we want the prose to say?
- how exactly will the upload component work?

and other such future problems. For now, lay a foundation for additional
work to build on.

Issue #409 Create bulk upload route and page