Add BulkUploaderFilePicker component and story
This component enables picking a file for bulk upload by either:

1. Dragging a file onto the drop area, or
2. Clicking the drop area to use the OS’s native file picker.

It currently lacks complete validation and any uploading. It also
pointedly is not appropriate for generic use. It can be adapted for
generic use, but that would require a level of configurability we can’t
justify without a second use case.

Issue #411 Create file-upload component