Show base fields on add data page
Add in the description of the base fields table that @reefdog wrote, and
build out the table with the data passed in from the loader.

For now, skip a dedicated loading state and instead just render an empty
list when the base fields haven't been fetched yet. If and when we
decide that loading is taking too long, we can tackle that at the time.

There are some very long base field short codes which would make the
first column unreasonably wide, so make them wrap (thanks @reefdog!). We
also need to review those base fields, which may be the better solution.

The API will soon be returning descriptions of base fields - although as
of this writing that change hasn't been deployed, and they're all
currently empty anyway - so update the type definition to reflect the
new field. We're okay with null values here, so this can be deployed
either before or after the base field descriptions API change.

Issue #418 Add base field description table to bulk upload instructions

Co-authored-by: Justin Reese <justin@justinreese.com>