Refresh bulk upload history on upload
Instead of loading the history once (via a loader), we want the bulk
uploads list to be defined in a way that gets updated when a refresh is
invoked.

We also want to pass the ability for a refresh to be invoked when a bulk
upload is created.

Issue #527 Refresh bulk upload history after upload completes