Add download attr to CSV template link
On the bulk upload page, we link users to a CSV template hosted by the
service repository. The service’s HTTP server gets to set the attachment
disposition header that determines if the file is downloaded to disk or
opened in-browser. Which is fine, and its prerogative! But we want *our*
users to definitely download the file.

The `download` attribute to the rescue! We’ll also take this opportunity
to set a value for the attribute, which will be the downloaded file’s
filename.