Add Organization List Panel
This commit adds the abstract component 'ListGrid', which is used to
implement the OrganizationListGrid and reimplement the exiting
ProposalListGrid.  'ListGrid' takes two props, an array 'items' of
generic types and a rendering function to render each item in items as a
react component. Following this, ListGridItems are now defined inside of
ListGrid files, and passed as props to the generic ListGrid.
Stylesheets for ListGridItems have been consolidated into one file and
are to be imported as needed.  This commit also adds the pdc sdk
typeOrganizationBundle and a query to access multiple Organizations,
with constant variables to decide count and pagination.  Finally,
[deprecated eslint rules have been removed from the
package.json](https://eslint.org/docs/latest/rules/comma-dangle)

Note: In the ListGrid component, the function ```scrollToActiveItem```
is used when the page renders, so the user knows "where they are".

If we just scrolled the active organizations into view, however, the
user might not realize there are more organizations "above" the active
one; thus, we scroll to the previous element.  If it doesn't exist --
i.e., if the first organization is active -- then we can noop since it's
visible by default anyway.

Note that we only run this on initial page render, not every time the
active organization changes.  This is because we only need to orient the
user on load; after that, their clicks run the show.

Issue #598 Create organization detail page