Add Organization Table
This commit adds the route '/organizations' and the corresponding
page OrganizationList and it's components. Furthermore, it abstracts
the ListTable into it's own component and refactors the current
ProposalListTable to use the new "ListTable" component. This is to
lay the infrastructure for when basefields are implemented into
the Organization schema. As they are not yet implemented, the
abstract ListComponent has the fieldNames as an optional prop,
which is used by the proposalListTable.

Fix Issue #596 Create Organization Table