Modify ProposalListGrid and Styles
This commit modifies the ProposalListGridItem class inside of the
ProposalListGrid item to rearrange the order of content, with proposal
name first, then organization name, then location. It also generalizes
the ListGridItem.css file further, into just a list-grid-item-title and
list-grid-item-detail.  Finally, in the case that the proposal title is
defaulting to the proposal-summary base field, it truncates the summary
to a set length

Fix Issue #602 modify proposal ui changes