Modify ProposalDetailPage
This commit modifies the proposal detail page to be titled by the
proposal name(or, if that is blank, a truncated proposal summary, or if
that is blank, a placeholder string), and moves the organization name to
a below field. It also removes the DataProvider button and functionality
from the page, along with the corresponding proposalId prop field.

Fix Issue #602 Implement UI changes for proposal detail page (in an org world)