Add Proposals table to Organization Detail Page
This commit adds a view for proposals inside of an organization detail
panel.  It does so by loading the proposal data(currently all proposal
data) in the OrganizationDetailLoader component and passing that to the
OrganizationDetail component, which then loads a ProposalListTable
component with the relevant data.

Fix Issue #612 Add Proposal table to Org page