Add the ListGridItem component
We previously began consolidating list grids with common styling and a
`ListGrid` parent component, but the individual grid items were still
defined in-place and not shared.

This commit introduces `ListGridItem`-related components (and stories),
and converts `ProposalListGrid` and `OrganizationListGrid` to use it.

It also introduces a new pattern: since none of the `ListGrid`-related
components are meant to be used individually outside of this tree, all
of them are defined in and exported from the same file (`ListGrid.tsx`).
Maybe we won’t like this! Maybe it’s bad! Maybe we decided earlier we
shouldn’t! But it reduces boilerplate for these very slim components,
and I may convert similar components (like `Dropdown`).

Also note the abundant use of `-webkit`-prefixing in the CSS properties
for limiting title text display to two lines; this is currently what’s
required to have cross-browser-compatibile line clamping.[1]

[1] https://css-tricks.com/line-clampin/

Issue #602 Implement UI changes for org-world proposal detail page

Co-authored-by: Harry Minksy <harry.minsky@gmail.com>