Change proposal details in sidebar
Replace the applicant name with the proposal name (or summary/fallback)
and move applicant name down to metadata area.

Not all proposals have names, so if absent, we show the proposal summary
instead. We have a utility method already for doing this with applicant
names, so this commit also abstracts out the core of that utility and
adds a variant for proposal name.

Two notes:

- As the data becomes more standardized, we should be able to stop
  cascading or falling back this way.
- I’m currently *not* truncating the value, even when it falls back to
  the potentinally Very Long `proposal_summary` field. I’m going to lean
  on CSS to truncate for us wherever it appears, and wait to see if we
  need to truncate for performance or other reasons later.

Issue #602 Implement UI changes for org-world proposal detail page

Co-authored-by: Harry Minksy <harry.minsky@gmail.com>