Add ProposalTable to OrganizationDetail
This commit adds the ProposalTable component
to the OrganizationDetail component, and sets up the OrganizationDetail
page to query the database for proposals related to the given organization,
using react state and effect hooks. It also adds a new method in the pdc-api.ts
file, UseProposalByOrganizationId, that accepts an organization id as a parameter
rather than a search query.

Issue #612 Add Proposals table to Organization Detail Page