Add OrganizationProposal Components
This commit adds a side panel to the organization detail page that
renders a related proposal. It introduces the OrganizationProposalLoader
and OrganizationProposalPanel components, which follow the convention of
the DataPlatformProvider component. OrganizationProposalPanel wraps the
CloseablePanel component, which ultimately wraps an instance of
ProposalListTable.  As it reuses the ProposalListTable but for an
organization context, it adds an optional prop to the ProposalListTable
component, an enum ProposalDetailDestinations, which is intended to
distinguish the endpoints of the row click, between a proposal detail
page and an organization detail panel.

Issue #613 Add Proposal Side Panel to Organization