Add DropdownMenuLinkDescription for dropdowns
Prior to this commit, we had a bit of an ad hoc / HTML-native way of
adding descriptions to dropdown menu items, customized just for our two
current navbar dropdowns. This felt unidiomatic both for React and for
our Dropdown component ergonomics.

Now, we have a `DropdownMenuLinkDescription` component that you can
optionally add to a `DropdownMenuLink`, alongside the existing text,
which renders some smaller explanatory text below the title.

Along the way, we also:

- Change the styling of these navbar items with descriptions so that
  they match items without descriptions. (h/t @slifty)
- Make sure menu link icons stay aligned with the “title” of the link,
  not centered by the description — and don’t let them shrink
- Update stories to match (and with more representative examples)

Issue #667 Change formatting of navbar dropdowns