Improve styling of fixed table in headless panel
When using a fixed table in a panel that had no panel header, the
table’s top border looked funky™.

This commit adds some overrides for that scenario, removing the top
border entirely.

Not directly related to the following issue, but noticed due to it.

Issue #688 Organization table isn't padded correctly