Add public data banner atop site
Reconfigure the app layout to support an optional banner atop the page.
For now, this banner is only used to alert public users that they may be
viewing limited data, and prompts them to log in. In future, we may
have additional uses for the banner.

In order to support this, the layout changes required some weirdly high
touches to the grid layout, such as moving padding off the `.App` shell
and onto individual main grid elements.

Issue #571 Add "you're seeing limited data" banner to public views