Relocate page security to router
We have multiple routes that point to the same page component, and in a
future change we will want to protect some of those routes differently
from others. (E.g., `/organizations/1` accessible to public users, but
`/organizations/1/proposals/2` inaccessible.)

To do this, we need to move securing them off the page component itself
to the router, where we can secure routes as needed. (We can still
secure individual components within pages using other methods.)

I am genuinely uncertain if there are significant downsides to this. The
primary reason for not doing this originally[1] was that we were not, at
the time, using `<RouteProvider>`, but we have (since?) begun using it.

[1] https://github.com/PhilanthropyDataCommons/front-end/pull/106#discussion_r1186521329

Issue #693 Secure pages at route definition