Eject react scripts
This commit ejects the react-scripts dependency from the project, using
the command 'npm run eject.' The only change made besides those made
automatically by the eject command is a manual update of the jest
package from v 26.3 to v 28.1.1, as per a dependency requirement from
the latest version of storybook.  v28.1.1 was chosen as it is the
current version used in the pdc/service repository.

Issue #692 Eject React Scripts