Take user to home page after login
Previously, we called the Proposals page “Dashboard”, which was useful
for user orientation but not really accurate.

Since we can’t be sure whether the user wants to go to proposals or orgs
after login, let’s bring them back to the landing page to choose.

Issue #561 Update actions on landing page for public view