Add ToggleSwitch components and stories
Adds an iOS-like toggle switch to use as a checkbox.

Includes an ESLint configuration change from requiring that a label
associate with its related input BOTH ways -- nesting the input within
the label AND matching the input's `id`  with the label's `htmlFor` --
to EITHER way. I think the "both" default is a bug, because the
documentation for this rule[1] states that either is correct.

[1] https://github.com/jsx-eslint/eslint-plugin-jsx-a11y/blob/main/docs/rules/label-has-associated-control.md

Issue #620 Design ability to filter proposals by "mine"