Use Preload on react-oidc
This commit updates the @axa-fr/react-oidc package to it's latest
version, 7.22.4 (NOTE: This version was released yesterday), to use it's
latest feature, user preload. Currently, the useOidcUser hook runs
async, which makes it a pain to couple with subsequent hooks. However,
[as added in the latest
version](https://github.com/AxaFrance/oidc-client/pull/1352), we can
explicitly set a configuration variable to await the loading of the user
data. Since we need to grapple the oidc user authentication id with the
authentication id in the database in order to get the DATABASE user id,
which is also done via an async hook, this seems like the correct
approach.

Issue #702 Get User's Id For Use in Dependent API Calls
seems to be the correct approach until