Add UserContext
This commit adds a new [Context
Provider](https://react.dev/reference/react/createContext) to the
front-end application with the goal of providing the User object that
corresponds to the OIDC user subject out from the database and
subsequently making that data available to any sub component. The
intention here is to minimize the number of API calls to the
pdc-service. This provider is layered under the current OIDC provider,
but is notably the first instance of a custom context provider in the
project.

Fix Issue #702 Get the user's ID for use in dependent API calls