Only show the user their own bulk uploads
We only want users to see their own bulk uploads. This is an intentional
front-end UX choice until we build out a fuller interface for admins to
see other people’s bulk uploads (if ever needed). It’s not a security
concern; the API handles making sure users only see what they’re
authorized to.

The `createdBy=me` value is a magic value that tells the API that the
scope should be “the user ID of the requesting user”.

Issue #697 Filter bulk uploads to current user