Correct vertical alignment of panel actions
We want most panel header children to vertically align at flex start,
that is, at the top of the header. (This way, for example, buttons in
top right corner of the panel stay there even when the title wraps.)

Switches, however, we want to sit a bit lower. For our current use, this
is easily accomplished by having them center-align; in the future, we’ll
probably want to just add some top padding so that they remain top-
aligned, but visually aligned with sibling buttons.

No issue.