Convert ApiBaseField to BaseField from SDK
In the first of several conversions, we change our own locally-defined
base field type, `ApiBaseField`, to the canonical `BaseField` type from
`@pdc/sdk`. Yay!

We also decorate the import with the `type` keyword when possible.
I.e., when using it as a type and not accessing any member attributes.

Issue #705 Use types from SDK when available

Co-authored-by: Justin Reese <justin@justinreese.com>