Convert ApiBulkUpload to BulkUpload from SDK
Remove `ApiBulkUpload`, use `BulkUpload`, and adapt to it. Additionally,
remove 'ApiBulkUploadsPostResponse' and 'ApiBulkUploadsPostRequest,'
replacing with 'WritableBulkUpload'

This require some changes to our datetime utilities, since now we’re
more natively passing in Date objects instead of strings. Along the way,
I realized we had some JSDoc errors in the utility function docs, which
are now fixed. (These were relevant to the type expansion, so it makes
sense to be included in this commit.)

Issue #705 Use types from SDK when available

Co-authored-by: Justin Reese <justin@justinreese.com>