Guard presigned post field values
We use Object.entries to build the headers of a presigned post without
checking values to ensure they are of the expected type (string or
Blob).

This was technically not type safe.  We initially had planned to address
this somehow in our sdk type definitions, but upon inspection the root
issue is the `Object.entries` signature, and so a type guard at the
point of entry iteration is the solution.

Issue #739 Remove ESLint overrides once SDK fixes any issues