Replace removed Web Vitals methods
Web Vitals v3 deprecated several methods and types, and v4 fully removed
them. This commit adapts our code to the breaking changes.

Note that the breaking removal of the `ReportHandler` type has been
complained about[1] and may return in deprecated form[2], but we should
go ahead and replace it anyway with the suggested inline typedef.

[1] https://github.com/GoogleChrome/web-vitals/issues/482
[2] https://github.com/GoogleChrome/web-vitals/pull/483

PR #745 Bump web-vitals from 3.5.2 to 4.0.0