Remove `WritablePresignedPostRequest` type fix
Prior to this change, we had to craft our own fixed
`WritablePresignedPostRequest` type to omit the `presignedPost` property
for use in generating a presigned post.

Now that the SDK has been updated to generate the correct
`WritablePresignedPostRequest`, we can use it directly.

At the same time, the SDK renamed `PresignedPostRequestPresignedPost` to
a more terse and accurate `PresignedPost`.

Issue #740 Replace FixedWritablePresignedPostRequest with version from
           SDK