Eject react scripts
This commit ejects the react-scripts package from the repository using
the built-in 'npm run eject' command, which is bundled with
create-react-app.  As the CRA package is no longer being maintained, we
want it out of the repository and replaced with vite. The produced
webpack config and scripts folders will be deleted in the subsequent
commit

Issue #720 migrate to Vite