Migrate to Vite
This exhaustive commit contains the work done by @slifty and myself(but
mostly @slifty) to migrate our build system from the now-deprecated
Create React App to Vite. Principle changes include bringing us to the
latest version of Typescript and Version 8.57 of Eslint. Notably, by
removing the react-scripts package, we now no longer extend the Create
React App default eslint rules. As such, two eslint disables in the
'pdc-api.ts' file have been removed for the time being, as they
reference rules we no longer use.

Another significant change is switching from jest to the 'vitest'
framework, which is the testing framework officially used by the project
Vite [1], and is designed to be natively compatible with vite.config.js
files. Note the additions of the vitest and jsdom packages.

It modifies Storybook to use the storybook/react-vite package,
as this is the storybook-reccomended plugin[2]. Furthermore, it removes
extrenous storybook plugins for CRA that are no longer needed.

It also disables the 'react/react-in-jsx-scope' rule from our
eslint, as it is used by our eslint-plugin-react package, but no longer
necessarry as of React 19 [3]. This removes the need for the 'import
React' statement above all components.

[1] https://vitest.dev/guide/#projects-using-vitest
[2] https://storybook.js.org/docs/get-started/react-vite
[3] https://legacy.reactjs.org/blog/2020/09/22/introducing-the-new-jsx-transform.html#eslint
Issue #720 Move To Vite

Co-authored-by: Daniel Schultz <slifty@gmail.com>