Fix BulkUploadListItem metadata reducer
This commit drops the initial value from the metadataElements reducer
function, along with the conditional statement. As per the MDN docs[1],
"if [no initial value is supplied,] the array element at index 0 is used
as the initial value." Thus we don't need to specify an initial value, as
we default to the first value, which, if it exists, is determined to be
not null by the filter function.

Issue #803 Remove initial value for BulkUploadListItem reducer function