Add trailing slash to Storybook link
When the app is built with the environment variable
VITE_SHOW_STORYBOOK=true, a link to Storybook is included in the
developers tab. The link works when hosted by Netlify, but DigitalOcean
App Platform is a bit pickier. Add the trailing slash, so that the
Storybook assets will load correctly.

Issue PhilanthropyDataCommons/deploy#118 Migrate to DigitalOcean Apps