Remove incorrect script documentation
The README entries were left over from create-react-app, and are no
longer accurate. More than that, this is a standard
JavaScript/TypeScript project, so the usual affordances are available:
`npm run` to see the options, `npm start`, `npm run build`, `npm test`,
and so on. I don't think we need project-specific documentation on
these.

Issue #720 Move To Vite