Add source handling for bulkuploads
This commit adds a method for providing a source for bulkuploads on the
front-end. Using the new useSystemSource hook, we query the pdc-service
for the FIRST created source (id:1), and use that to provide a sourceId
value. Furthermore, it updates date fields in bulkuploads to be strings,
and ids to be uuids, from Date and integer types, repsectively. This
follows the change made in the pdc service.