Add source label to proposal version
This commit adds the rendering for a given proposal version's Source,
following the same pattern as we currently use for applicant and tax ID
(the Panel Tag component). It also modifies the ClosablePanel component
to take an optional tags prop, which is used to render the source in
the proposal-changemaker sidepanel.

Issue 913 Display Source On Proposal Versions