Group react more aggressively
It seems to me there is very little downside in trying to group our
dependencies -- we know the pool of dependencies we're considering, and
if somehow we get a grouping wrong we can adjust later.

There are some react bumps that depend on other items right now, and the
PRs cannot be merged independently.  This groups all react related
changes into one place.