Add in minimal vue application and set react code aside
This commit introduces a minimal VueJs application to the repository,
preserving as much editing and linting configuration as possible, and
sets the existing react code aside, to be used as a reference for
rebuilding the front-end in Vue. Thus far, no dependencies have
been removed, but a slew of vue related dependencies have been
added, namely the vue package itself and the official vue
eslint and typescript configurations. Since we are no longer
developing the react codebase, CI/CD is now only pointing at
the new Vue files. Subsequent commits will involve the rebuilding
of each of our existing react components in the vue-est way possible,
and re-establishing our routing and authentication system.

Issue #1009 Migrate front-end to Vue