Rebuild landing page in vue, migrate all components in-use
This commit contains the migration of all components that are used and
rendered currently in the react application.  We'll note that, in
keeping with Vue syntax, where we would previously refer to route-level
components as 'pages,' we now use 'views'. Furthermore, in perhaps a
less vue-y fashion, we are still keeping the components split up at the
level they were in the react application (i.e separate Dropwdown and
Panel sub-components). We can always refactor this setup later, but for
the time being this allows us to recreate the functionality of our react
application quite closely.

A somewhat significant change in syntax/design from react to vue seen in
this commit is the structure of templating 'child' components.  In react
we explicitly past 'children' props of type React.ReactNode to our
components, to be rendered within them.  Vue's suggested syntax for this
is to use slots, which are not explicitly typed, but are implicitly
assumed to be components.

A smaller change was also made in the reconstruction of the EmailLink
component, which previously used a home-brewed solution to parsing
values into email-acceptable query strings. Instead, we now use the
qs[1] package, which is a widely used package for accomplishing just
such a task.

[1]https://www.npmjs.com/package/qs

Issue #1009 Migrate front-end to Vue