Add keycloak auth integration
This commit introduces authentication handling for the application using
the "@dsb/vue-keycloak" package. While the package is not widely
used, it is actively maintained (most recent release Feb 19th 2025 [1]),
and importantly, is a Vue3 wrapper AROUND the keycloak-js official
sdk. Which is to say, it contains no additional authentication logic
beyond that which is needed to make the keycloak sdk vue-compatible.
Similarly to the existing @axa-fr/react-oidc package we used in
the old application, we expose a reactive component which is
populated with authentication data and can be used throughout the
application. Furthermore, it is compatible with the vue-router
package.

Issue #1009 Rebuild front-end in Vue

[1] https://github.com/dsb-norge/vue-keycloak-js/releases/tag/v3.0.2