Add custom pdc-api hook
This commit 'migrates' our generic api hook for fetching resources from
the pdc-api, replacing react's reactivity and state-based functions with
those from vue. Also, our new auth package sadly does not provide an
authorized fetch function, so we are instead using it's getToken hook
and adding that to our fetch header. In trying to recreate the exisiting
react function as closely as possible, there is likely room here for
making things more vue-y in the future. But having tested locally it
seems to function as expected.

Similar to react's useEffect, we would want to call the fetchData
function as the callback to vue's onMounted lifecylce hook.

Issue #1009 Rebuild frontend in Vue