Rebuild Storybook components in vue
This commit migrates our storybook setup from react to vue, using the
vue3/vite storybook package. Most functionality is preserved from react
in storybook, with the only major change being in templating structure:
whereas before we passed child components as jsx props, since Vue
instead promotes using slots, we set the children of our stories using
the template syntax.  Also, there is not native vue-router support, so
we have to mock the vue-router package around our stories in order to
use any components that rely on vue-router package-specific components.

The withVueRouter function is modeled after the same function found
here[1] in a vue-router-storybook plugin. Since we have no demonstrated
need for actually mocking the functionality of the router, it seemed the
right decision to just include the bare minimum code needed to render
the child components, rather than adding another dependency (that is
both not actively maintained and depends on an earlier version of
storybook).

Issue #1009 Migrate Front-end to vue

[1] https://github.com/NickMcBurney/storybook-vue3-router