Remove unused packages
This commit removes all currently unused packages from the repository,
which include react, react-specific linting, and a few utility packages
which were used for app functionality that may not be used moving
forward. Notably, we still need @vue/plugin-react to bring in necessary
assets for @vue/eslint-config-airbnb-with-typescript.  We are in the
process of upgrading to eslint 9 and moving away from the airbnb config,
as it is now an archived repository.

Issue #1033 Remove unused packages