Remove airbnb eslint configuration
Following the same move made in the pdc-service repo[1], this commit
removes the airbnb eslint configuration (and the @vitejs/plugin-react
repo, which was seemingly a dependancy for the vueAirbnbTypescript
plugin we were using) from the project, as it is no longer being
maintained and is forcing us to stay pinned to a now-deprecated version
of ESLint. As was done in the service, we're adding one disable for a
rule which requires arrow functions to have an `await` in order to be
allowed to be `async`, which we deemed to be silly.

[1] https://github.com/PhilanthropyDataCommons/service/pull/1703

Issue #1032 Upgrade to ESLint 9