Upgrade to ESLint9
This commit upgrades the repository to ESLint9, which all of our
existing eslint-configuration packages pleasently support (in their
latest version, notably this brings us to the LTS for the
typescript-eslint plugin and typescript itself, which is a plus).  It
removes our old json configuration file for the new standardized flat
config js file. The flat config was partially structured off of the
pdc-service repo, as well as the template for eslint9 with vue provided
on the eslint-vue-plugin website[1].

[1] https://eslint.vuejs.org/user-guide/#example-configuration-with-typescript-eslint-and-prettier

Issue #1032 Upgrade to ESLint 9