Add Love eslint config
This commit adds in the Love eslint configuration, which we've
determined to be a suitable replacement for the now removed (and
deprecated) airbnb eslint config. It disables all rules that were
causing conflicts so that they can be re-enabled individually commit by
commit. The only changes made to accomodate love in this commit are the
addition of the '.vue' file type to our parser options, and the
refactoring of an index.tsx file (which very much should not have been
in the repo, now that we are not using tsx. thank you love!) to a
regular typescript file.

Issue #1056 Move to Love ESLint configuration