Re-enable @typescript-eslint/explicit-function-return-type
This commit adds explicit function return types to all functions in the front-end, and
re-enables the corresponding eslint rule.

A note about the pdc-api usePdcApi function:

In order to type the return value using our generic T parameter, we had
to also change the way we typed the data variable. The previous method
of using an interface with the ref function was actually returning a
heavily overloaded type Normally the ref<T> interface serves the
function perfectly well and is a cleaner write, but since we are dealing
with typing the return values of the refs, it seems reasonable to
explicitly type the variable with a narrower type.

This commit has also brought to my attention that we may need to rethink
this function -- it's supposed to recreate the functionality of our old
pdc-api custom hook that used react hooks[1], but we haven't done much
research into the proper 'vue' way to fetch and refresh data
re-actively.  We should make an issue to research this further,
especially considering we intend to introduce editing capabilities
between the front-end and service.

[1] https://github.com/PhilanthropyDataCommons/front-end/blob
/c6b46383b6e470c461880fae4565fedd69c5b9bc/src/pdc-api.ts

Issue #1056 Move to Love ESLint configuration