Re-enable @typescript-eslint/no-unnecessary-condition
This commit re-enables another love rule,
`@typescript-eslint/no-unncessary-condition`, which as the name suggests
flags any checks that are redundant against our typing. The only reason
that I'm adding a commit message here is to note that, in keeping with
the old front-end application, I opted to treat the LOG_LEVEL
envrionment variable as optional rather than remove the nullish
check. In theory, we could also decide to just enforce the variable, but
that should involve a separate determination process.

Issue #1056 Move to Love Eslint configuration