Re-enable @typescript-eslint/no-unsafe-type-assertion
This commit removes all unsafe type assertions and reenables the
corresponding ESLint rule, save for one explicit disable in the
`pdc-api` hook. The assumption here is that IF we are communicating with
a valid pdc service instance, and the request is successful, that the
response should be valid JSON and of the expected type. We could
introduce dynamic type-guarding in the future, but for the time being it
seems reasonable to put the burden of type-checking on the pdc service,
so long as we verify the success of the request.

Issue #1056 Move to Love ESLint configuration