Re-enable @typescript-eslint/no-unnecessary-type-parameters
This commit ensures that all type parameters are used in functions,
which currently was only relevant to our generic custom pdc-api
hook. Helpfully, it exposed an issue in our typing:

Before, the function typed the returned data as 'Ref', which without an
interface is processed as 'Ref<any,any>', meaning that while the
explicitly typed return value for the 'useBaseFields' hook was still
typing the return data as 'Ref<BaseField[]|null>', the output of `return
usePdcApi<BaseField[]>('/baseFields');` on Line 60 was typing it's
return as 'Ref<any,any>'.

Something that came up while fixing this: in order to type the return
value using our generic T parameter, we had to also change the way we
typed the data variable. The previous method of using an interface with
the ref function was actually returning a heavily overloaded type

`(ref<T>(value: T): [T] extends [Ref<any, any>] ?  IfAny<T, Ref<T, T>, T>
: Ref<UnwrapRef<T>, T | UnwrapRef<T>>).`

It seems like we don't need the
full scope of that typing for our purposes, and in fact if the returned
value is anything BESIDES Ref<T|null> then it seems like something would
have gone wrong with our fetch, which should be caught by the try catch
block. Normally the ref<T> interface serves the function perfectly well
and is a cleaner write, but since we are dealing with typing the return
values of the refs, it seems reasonable to explicitly type the variable
with a narrower type (which would be a no-no if we were doing so after
the fact)

This commit has also brought to my attention that we may need to rethink
this function -- it's supposed to recreate the functionality of our old
pdc-api custom hook that used react hooks[1], but we haven't done much
research into the proper 'vue' way to fetch and refresh data re-actively.
We should make an issue to research this further, especially considering
we intend to introduce editing capabilities between the front-end and
service.

[1] https://github.com/PhilanthropyDataCommons/front-end/blob
/c6b46383b6e470c461880fae4565fedd69c5b9bc/src/pdc-api.ts

Issue #1056 Move to Love ESLint config